fix: answer, comment, post BaseEntity 상속 분리, nullable = true
