+++ src/main/java/com/hyetaekon/hyetaekon/answer/entity/Answer.java
@@ -5,6 +5,10 @@ import com.hyetaekon.hyetaekon.post.entity.Post;
 import com.hyetaekon.hyetaekon.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import java.time.LocalDateTime;
 
 
 @Entity
@@ -14,7 +18,8 @@ import lombok.*;
 @NoArgsConstructor
 @AllArgsConstructor
 @Table(name = "answer")
-public class Answer extends BaseEntity {
+@EntityListeners(AuditingEntityListener.class)
+public class Answer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id; // 답변 ID
@@ -33,10 +38,30 @@ public class Answer extends BaseEntity {
     @Column(name = "selected", nullable = false)
     private boolean selected; // 채택 여부
 
+    @CreatedDate
+    @Column(name = "created_at", nullable = false, updatable = false)
+    private LocalDateTime createdAt;
+
+    @Column(name = "deleted_at")
+    private LocalDateTime deletedAt;
+
+    @Column(name = "suspend_at")
+    private LocalDateTime suspendAt;
+
+    // 삭제 처리
+    public void delete() {
+        this.deletedAt = LocalDateTime.now();
+    }
+
+    // 정지 처리
+    public void suspend() {
+        this.suspendAt = LocalDateTime.now();
+    }
+
     public String getDisplayContent() {
-        if (getDeletedAt() != null) {
+        if (this.deletedAt != null) {
             return "삭제된 답변입니다.";
-        } else if (getSuspendAt() != null) {
+        } else if (this.suspendAt != null) {
             return "관리자에 의해 삭제된 답변입니다.";
         }
         return content;
