+++ src/main/java/com/hyetaekon/hyetaekon/comment/entity/Comment.java
@@ -5,6 +5,8 @@ import com.hyetaekon.hyetaekon.post.entity.Post;
 import com.hyetaekon.hyetaekon.user.entity.User;
 import jakarta.persistence.*;
 import lombok.*;
+import org.springframework.data.annotation.CreatedDate;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
 
 import java.time.LocalDateTime;
 
@@ -15,7 +17,8 @@ import java.time.LocalDateTime;
 @NoArgsConstructor
 @AllArgsConstructor
 @Table(name = "comments")
-public class Comment extends BaseEntity {
+@EntityListeners(AuditingEntityListener.class)
+public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -33,10 +36,30 @@ public class Comment extends BaseEntity {
     @Column(nullable = false, length = 1000)
     private String content;
 
+    @CreatedDate
+    @Column(name = "created_at", nullable = false, updatable = false)
+    private LocalDateTime createdAt;
+
+    @Column(name = "deleted_at")
+    private LocalDateTime deletedAt;
+
+    @Column(name = "suspend_at")
+    private LocalDateTime suspendAt;
+
+    // 삭제 처리
+    public void delete() {
+        this.deletedAt = LocalDateTime.now();
+    }
+
+    // 정지 처리
+    public void suspend() {
+        this.suspendAt = LocalDateTime.now();
+    }
+
     public String getDisplayContent() {
-        if (getDeletedAt() != null) {
+        if (this.deletedAt != null) {
             return "삭제된 댓글입니다.";
-        } else if (getSuspendAt() != null) {
+        } else if (this.suspendAt != null) {
             return "관리자에 의해 삭제된 댓글입니다.";
         }
         return content;
